Testing

+++ backend/src/main/java/com/turismea/service/CityService.java
@@ -5,6 +5,8 @@ import com.turismea.model.City;
 import com.turismea.repository.CityRepository;
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
+
 @Service
 public class CityService {
     private final CityRepository cityRepository;
@@ -14,7 +16,15 @@ public class CityService {
     }
 
     public void deleteCity(Long cityId) {
+        if(cityId == null) {
+            throw new IllegalArgumentException();
+        }
         City city = cityRepository.findById(cityId).orElseThrow(() -> new CityNotFoundException(cityId));
         cityRepository.delete(city);
     }
+
+    public Optional<City> findByName(String city) {
+        return Optional.ofNullable(cityRepository.findByName(city));
+    }
+
 }
